+++ CHANGELOG.md
@@ -9,6 +9,7 @@
     - `BrowserController` - Wraps a single browser instance and maintains its state.
     - `BrowserPlugin` - Manages the browser automation framework, and basically acts as a factory for controllers.
 - Browser rotation with a maximum number of pages opened per browser.
+- Add emit persist state event to event manager
 
 ## [0.0.4](../../releases/tag/v0.0.4) - 2024-05-30